+++ integration-tests/gatsby-cli/__tests__/build.js
@@ -12,9 +12,8 @@ describe(`gatsby build`, () => {
   it(`creates a built gatsby site`, () => {
     const [code, logs] = GatsbyCLI.from(cwd).invoke(`build`)
 
-    logs.should.contain(
-      `success open and validate gatsby-configs, load plugins`
-    )
+    logs.should.contain(`success load gatsby config`)
+    logs.should.contain(`success load plugins`)
     logs.should.contain(`success onPreInit`)
     logs.should.contain(`success initialize cache`)
     logs.should.contain(`success copy gatsby files`)