+++ packages/gatsby-plugin-gatsby-cloud/src/utils/use-poll-for-node-manifest/types.ts
@@ -12,6 +12,7 @@ type FoundPageBy =
   | `ownerNodeId`
   | `filesystem-route-api`
   | `context.id`
+  | `context.slug`
   | `queryTracking`
   | `none`