+++ packages/gatsby-sharp/.babelrc.js
@@ -1,4 +1,3 @@
-console.log('hi there');
 module.exports = {
   "presets": [["babel-preset-gatsby-package"]],
   "plugins": ["babel-plugin-replace-ts-export-assignment"]