+++ packages/gatsby/src/schema/__tests__/build-schema.js
@@ -36,6 +36,7 @@ jest.mock(`gatsby-cli/lib/reporter`, () => {
     info: jest.fn(),
     warn: jest.fn(),
     error: jest.fn(),
+    panic: jest.fn(console.error),
     activityTimer: () => {
       return {
         start: jest.fn(),
@@ -1392,6 +1393,68 @@ describe(`Build schema`, () => {
           sharedField: String
           originalField: String
           newField: String
+        }"
+      `)
+    })
+
+    it(`Can reference derived types when merging types`, async () => {
+      createTypes(gql`
+        # create initial type composer
+        type TypeCreatedBySourcePlugin implements Node {
+          id: ID!
+          someField: String
+        }
+      `)
+      createTypes([
+        buildInterfaceType({
+          name: `SharedInterface`,
+          fields: {
+            id: `ID!`,
+            child_items: {
+              type: `[SharedInterface]`,
+              args: {
+                // referencing derived type
+                sort: `SharedInterfaceSortInput`,
+              },
+            },
+          },
+          interfaces: [`Node`],
+        }),
+        buildObjectType({
+          name: `TypeCreatedBySourcePlugin`,
+          fields: {
+            id: `ID!`,
+            child_items: {
+              type: `[SharedInterface]`,
+              args: {
+                sort: `SharedInterfaceSortInput`,
+              },
+              resolve: (_, args) => [],
+            },
+          },
+          interfaces: [`Node`, `SharedInterface`],
+        }),
+      ])
+
+      // implicit assertion is that building schema doesn't throw in the process
+      const schema = await buildSchema()
+      expect(printType(schema.getType(`TypeCreatedBySourcePlugin`)))
+        .toMatchInlineSnapshot(`
+        "type TypeCreatedBySourcePlugin implements Node & SharedInterface {
+          id: ID!
+          someField: String
+          child_items(sort: SharedInterfaceSortInput): [SharedInterface]
+          parent: Node
+          children: [Node!]!
+          internal: Internal!
+        }"
+      `)
+
+      expect(printType(schema.getType(`SharedInterfaceSortInput`)))
+        .toMatchInlineSnapshot(`
+        "input SharedInterfaceSortInput {
+          fields: [SharedInterfaceFieldsEnum]
+          order: [SortOrderEnum] = [ASC]
         }"
       `)
     })